refactor: Clean up repository root directory
- Move Python scripts to tools/ directory
- Move test files to tests/ directory
- Move documentation to docs/ directory
- Move slash commands to .agent-os/commands/
- Archive temporary and backup files
- Remove build artifacts and coverage files

Root directory now contains only essential files:
- pyproject.toml, README.md, LICENSE
- uv.toml, uv.lock
- Makefile, CLAUDE.md
- Git configuration files
