+++ .agent-os/commands/slash_commands.py
