+++ archive/MANIFEST.in
