+++ docs/MANDATORY_SLASH_COMMAND_ECOSYSTEM.md
